feat: 존재하지 않는 예약 삭제시 예외 발생

+++ src/main/java/roomescape/service/ReservationService.java
@@ -55,6 +55,10 @@ public class ReservationService {
     }
 
     public void deleteReservation(final long id) {
+        if (reservationRepository.existsById(id)) {
+            throw new RoomEscapeException(ErrorCode.RESERVATION_NOT_FOUND);
+        }
+        
         reservationRepository.deleteById(id);
     }
